requirements.txt
python-telegram-bot==21.6
openai>=1.30.0
gspread>=6.1.2
google-auth>=2.34.0
APScheduler>=3.10.4
pytz>=2024.1
python-dotenv>=1.0.1
